// src/app/shared/types/api-types/bire-localisation-item-dto.interface.ts
export interface BireLocalisationItemDTO {
  localisationCode?: string;
  familyCode?: string;
  structureType?: string;
  variantCode?: string;
  chapter?: string;
  sheet?: string;
  marks?: string;
  section?: string;
  subject?: string;
  versionNumber?: string;
  localisationDescription?: string;
}